# specs/001-mvp-closed-loop/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: PPTagent MVP API
  version: 0.1.0-mvp
  description: |
    REST API for PPTagent MVP 业务闭环（生成 / 知识库 / Agent 进化）。
    与 WebSocket 事件契约（events.yaml）配合使用，错误码遵循 RFC 7807。
servers:
  - url: http://localhost:8000/api/v1
    description: Local dev
  - url: https://api.pptagent.example.com/api/v1
    description: Production (placeholder)
security:
  - ApiKeyAuth: []
paths:
...
  /generations:
    post:
      summary: 一句话生成 PPT
      operationId: createGeneration
      tags: [Generations]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGenerationRequest'
      responses:
        '202':
          description: 已入队（含 queue_position）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenerationTask'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /generations/{task_id}:
    parameters:
      - in: path
        name: task_id
        required: true
        schema: { type: string, format: uuid }
    get:
      summary: 查询任务状态
      operationId: getGeneration
      tags: [Generations]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenerationTask'
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      summary: 取消任务（FR-003）
      operationId: cancelGeneration
      tags: [Generations]
      responses:
        '204': { description: 已停止 }
        '409': { $ref: '#/components/responses/Conflict' }
...
  /samples/batch:
    post:
      summary: 批量上传样本（FR-006）
      operationId: batchUploadSamples
      tags: [Samples]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                files:
                  type: array
                  items: { type: string, format: binary }
                  maxItems: 20
      responses:
        '202':
          description: 已入队解析
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Sample' }
        '413': { $ref: '#/components/responses/PayloadTooLarge' }
  /samples/{sample_id}:
    parameters:
      - in: path
        name: sample_id
        required: true
        schema: { type: string, format: uuid }
    get:
      summary: 样本详情（含 PII 摘要 + 解析结果）
      operationId: getSample
      tags: [Samples]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SampleDetail' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      summary: 删除单一样本（FR-007）
      operationId: deleteSample
      tags: [Samples]
      responses:
        '204': { description: OK }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      summary: 强制重新解析（FR-007 失败重试）
      operationId: reparseSample
      tags: [Samples]
      responses:
        '202': { description: 已入队 }
        '409': { $ref: '#/components/responses/Conflict' }

  # ─── US6: 素材库 (Material Library) ───
...
  /preferences/{preference_id}:
    parameters:
      - in: path
        name: preference_id
        required: true
        schema: { type: string, pattern: '^P-\d{3,}$' }
    get:
      summary: 偏好详情（含完整来源链）
      operationId: getPreference
      tags: [Preferences]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Preference' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      summary: 删除偏好（FR-013）
      operationId: deletePreference
      tags: [Preferences]
      responses:
        '204': { description: OK }
        '404': { $ref: '#/components/responses/NotFound' }

  # ─── Data Lifecycle ───
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: 形如 `Bearer <key>`（最小权限 IV + 速率限制）
...
  responses:
    BadRequest:
      description: 请求参数错误
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Unauthorized:
      description: 未认证
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NotFound:
      description: 资源不存在
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Conflict:
      description: 状态冲突（如取消已完成任务）
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    PayloadTooLarge:
      description: 单文件 > 50MB 或批量 > 20 份
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    RateLimited:
      description: 并发超限 / 配额耗尽
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
    CreateGenerationRequest:
      type: object
      required: [prompt]
      properties:
        prompt:
          type: string
          minLength: 1
          maxLength: 1000
          description: 一句话需求（FR-001）
        referenced_sample_ids:
          type: array
          items: { type: string, format: uuid }
          description: 可选，指定本次任务使用的样本子集
    GenerationTask:
      type: object
      properties:
        id: { type: string, format: uuid }
        prompt: { type: string }
        status:
          type: string
          enum: [queued, running, success, failed, cancelled, archived]
        current_stage:
          type: string
          enum: [outline, points, svg, pptx, null]
        queue_position:
          type: integer
          nullable: true
          description: FR-029 排队位置
        result_pptx_url:
          type: string
          format: uri
          nullable: true
        style_fit_score:
          $ref: '#/components/schemas/StyleFitScore'
        token_consumed: { type: integer }
        estimated_tokens: { type: integer, nullable: true }
        estimated_seconds: { type: integer, nullable: true }
        created_at: { type: string, format: date-time }
        expires_at: { type: string, format: date-time, nullable: true }
...
    Sample:
      type: object
      properties:
        id: { type: string, format: uuid }
        file_name: { type: string }
        file_type:
          type: string
          enum: [pptx, pdf, docx]
        parse_status:
          type: string
          enum: [pending, parsing, parsed, failed]
        parse_page_count: { type: integer, nullable: true }
        pii_summary: { $ref: '#/components/schemas/PIISummary' }
        uploaded_at: { type: string, format: date-time }
        parsed_at: { type: string, format: date-time, nullable: true }
        error_message: { type: string, nullable: true }
    SampleDetail:
      type: object
      description: 单样本详情（含 PII 处置摘要 + 解析结果）
      properties:
        id: { type: string, format: uuid }
        file_name: { type: string }
        file_type: { type: string, enum: [pptx, pdf, docx] }
        file_size_bytes: { type: integer, description: '原始文件大小（FR-006 ≤ 50MB）' }
        file_hash: { type: string, description: 'SHA-256，FR-010 去重' }
        parse_status: { type: string, enum: [pending, parsing, parsed, failed] }
        parse_page_count: { type: integer, nullable: true }
        pii_summary: { $ref: '#/components/schemas/PIISummary' }
        parse_result:
          description: 解析结构化结果（仅 parse_status=parsed 时返回）
          allOf:
            - $ref: '#/components/schemas/ParseResult'
        uploaded_at: { type: string, format: date-time }
        parsed_at: { type: string, format: date-time, nullable: true }
        error_message: { type: string, nullable: true }
    ParseResult:
      type: object
      properties:
        version: { type: string, description: 解析器版本（依赖锁定） }
        structure:
          type: object
          description: '页/版式/文本块的层级 JSON（PG 结构化摘要）'
          additionalProperties: true
        parsed_at: { type: string, format: date-time }
        page_count: { type: integer }
    PIISummary:
      type: object
      properties:
        total_hits: { type: integer }
        hits_by_field:
          type: object
          additionalProperties: { type: integer }
          example: { phone: 2, email: 1, id_card: 0, customer_name: 3 }
        fields_replaced: { type: integer }
    Preference:
      type: object
      properties:
        id: { type: string, pattern: '^P-\\d{3,}$', example: P-007 }
        rule_text: { type: string }
        applies_to:
          type: string
          enum: [cover, toc, body, closing, all]
        apply_count: { type: integer }
        ignore_count: { type: integer }
        last_applied_at: { type: string, format: date-time, nullable: true }
        is_active: { type: boolean }
        source_chains:
          type: array
          items: { $ref: '#/components/schemas/SourceChain' }
    SourceChain:
      type: object
      properties:
        type: { type: string, enum: [sample, generation, manual] }
        ref_id: { type: string, format: uuid }
        snippet: { type: string, maxLength: 500 }
        created_at: { type: string, format: date-time }
    StyleFitScore:
      type: object
      properties:
        layout: { type: number, minimum: 0, maximum: 1 }
        palette: { type: number, minimum: 0, maximum: 1 }
        font: { type: number, minimum: 0, maximum: 1 }
        overall_hit: { type: boolean, description: FR-028 三层均 ≥ 80% 时为 true }
...
    Problem:
      type: object
      description: RFC 7807
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string, format: uri }
        code: { type: string, description: PPTagent 自定义错误码，详见 error-codes.yaml }
        request_id: { type: string, format: uuid, description: 关联 X-Request-Id }
        trace_id: { type: string, description: OTel trace ID }
        errors:
          type: array
          description: 字段级错误（422 验证失败时填充）
          items:
            type: object
            properties:
              field: { type: string }
              code: { type: string }
              message: { type: string }

    # ─── Pagination ───
...